refactor: require mutable access via monitored_utxos_mut()
